Keeping the exit code of jobs
If a job in the list fails the build command still returns 0 as if nothing happens. This might not be the correct behaviour. In our example we have a build job that builds the assets but if the asset building fails then the builder still returns 0. Which is not true.